Updated example watcher_demo.py and readme
